+++ app/lib/pages/settings/integrations_page.dart
@@ -7,8 +7,8 @@ import 'package:omi/widgets/shimmer_with_timeout.dart';
 import 'package:omi/pages/apps/add_app.dart';
 import 'package:omi/pages/settings/apple_health_detail_page.dart';
 import 'package:omi/providers/integration_provider.dart';
-import 'package:omi/services/apple_health_service.dart';
-import 'package:omi/services/google_calendar_service.dart';
+import 'package:omi/services/integrations/apple_health_service.dart';
+import 'package:omi/services/integrations/google_calendar_service.dart';
 import 'package:omi/utils/l10n_extensions.dart';
 import 'package:omi/utils/logger.dart';
 import 'package:omi/utils/other/temp.dart';
