+++ app/lib/services/integrations/apple_health_service.dart
