+++ app/lib/services/integrations/apple_reminders_sync_service.dart
@@ -1,7 +1,7 @@
 import 'package:omi/utils/platform/platform_manager.dart';
 import 'package:omi/backend/http/api/action_items.dart';
 import 'package:omi/backend/schema/action_item.dart';
-import 'package:omi/services/apple_reminders_service.dart';
+import 'package:omi/services/integrations/apple_reminders_service.dart';
 import 'package:omi/utils/logger.dart';
 
 /// Orchestrates bidirectional Apple Reminders sync on foreground resume.
